+++ src/main/java/com/academy/project/hotelsmanagementsystem/repository/BookingRepository.java
@@ -3,5 +3,5 @@ package com.academy.project.hotelsmanagementsystem.repository;
 import com.academy.project.hotelsmanagementsystem.entity.BookingEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookingRepository extends JpaRepository<BookingEntity,Integer> {
+public interface BookingRepository extends JpaRepository<BookingEntity,Long> {
 }
