+++ src/main/java/com/academy/project/hotelsmanagementsystem/repository/HotelRepository.java
@@ -3,5 +3,5 @@ package com.academy.project.hotelsmanagementsystem.repository;
 import com.academy.project.hotelsmanagementsystem.entity.HotelEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HotelRepository extends JpaRepository<HotelEntity,Integer> {
+public interface HotelRepository extends JpaRepository<HotelEntity,Long> {
 }
