+++ src/main/java/com/academy/project/hotelsmanagementsystem/repository/RoleRepository.java
@@ -3,5 +3,5 @@ package com.academy.project.hotelsmanagementsystem.repository;
 import com.academy.project.hotelsmanagementsystem.entity.RoleEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleRepository extends JpaRepository<RoleEntity,Integer> {
+public interface RoleRepository extends JpaRepository<RoleEntity,Long> {
 }
