+++ src/main/java/com/academy/project/hotelsmanagementsystem/repository/RoomBookedRepository.java
@@ -3,5 +3,5 @@ package com.academy.project.hotelsmanagementsystem.repository;
 import com.academy.project.hotelsmanagementsystem.entity.RoomBookedEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoomBookedRepository extends JpaRepository<RoomBookedEntity,Integer> {
+public interface RoomBookedRepository extends JpaRepository<RoomBookedEntity,Long> {
 }
