+++ src/main/java/com/academy/project/hotelsmanagementsystem/repository/RoomRepository.java
@@ -3,5 +3,5 @@ package com.academy.project.hotelsmanagementsystem.repository;
 import com.academy.project.hotelsmanagementsystem.entity.RoomEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoomRepository extends JpaRepository<RoomEntity,Integer> {
+public interface RoomRepository extends JpaRepository<RoomEntity,Long> {
 }
