+++ src/main/java/com/academy/project/hotelsmanagementsystem/repository/RoomTypeRepository.java
@@ -3,5 +3,5 @@ package com.academy.project.hotelsmanagementsystem.repository;
 import com.academy.project.hotelsmanagementsystem.entity.RoomTypeEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoomTypeRepository extends JpaRepository<RoomTypeEntity,Integer> {
+public interface RoomTypeRepository extends JpaRepository<RoomTypeEntity,Long> {
 }
